requirements.txt
numpy==1.19.3
PyAutoGUI==0.9.52
Pillow==8.0.1
opencv-contrib-python==4.4.0.44